+++ threedi_modelchecker/migrations/versions/0200_initial.py
@@ -18,6 +18,7 @@ down_revision = None
 branch_labels = None
 depends_on = None
 
+
 existing_tables = []
 
 
@@ -28,11 +29,6 @@ def _get_existing_tables(inspector):
     existing_tables = inspector.get_table_names()
 
 
-def _get_column_lookup(inspector, table_name):
-    columns = inspector.get_columns(table_name)
-    return {c["name"]: c for c in columns}
-
-
 def create_table_if_not_exists(table_name, *args, **kwargs):
     """Create a table if it is not in the global 'existing_tables'"""
     if table_name in existing_tables:
@@ -41,36 +37,28 @@ def create_table_if_not_exists(table_name, *args, **kwargs):
         return op.create_table(table_name, *args, **kwargs)
 
 
-def drop_field_if_exists(inspector, table_name, field_name):
-    if field_name not in _get_column_lookup(inspector, table_name):
-        return  # field doesn't exist: do nothing
+def upgrade_173():
+    """This implements the latest migration from the old stack:
 
-    with op.batch_alter_table(table_name) as batch_op:
-        batch_op.drop_column(field_name)
-
-
-def make_field_nullable(inspector, table_name, field_name):
-    if _get_column_lookup(inspector, table_name)[field_name]["nullable"]:
-        return  # column is nullable: do nothing
-
-    with op.batch_alter_table(table_name) as batch_op:
-        batch_op.alter_column(field_name, nullable=True)
-
-
-def upgrade_173(inspector):
+    0172_auto__del_v2initialwaterlevel__del_field_v2orifice_max_capacity__del_f
+    """
     if "v2_initial_waterlevel" not in existing_tables:
-        return  # skip this migration
+        return  # skip this migration (probably already done)
 
-    # this duplicates the last migration from the old stack:
-    # 0172_auto__del_v2initialwaterlevel__del_field_v2orifice_max_capacity__del_f
     op.drop_table("v2_initial_waterlevel")
 
-    drop_field_if_exists(inspector, "v2_orifice", "max_capacity")
-    drop_field_if_exists(inspector, "v2_impervious_surface", "function")
-    drop_field_if_exists(inspector, "v2_pipe", "pipe_quality")
-    drop_field_if_exists(inspector, "v2_orifice", "max_capacity")
-    make_field_nullable(inspector, "v2_culvert", "discharge_coefficient_negative")
-    make_field_nullable(inspector, "v2_culvert", "discharge_coefficient_positive")
+    with op.batch_alter_table("v2_orifice") as batch_op:
+        batch_op.drop_column("max_capacity")
+
+    with op.batch_alter_table("v2_impervious_surface") as batch_op:
+        batch_op.drop_column("function")
+
+    with op.batch_alter_table("v2_pipe") as batch_op:
+        batch_op.drop_column("pipe_quality")
+
+    with op.batch_alter_table("v2_culvert") as batch_op:
+        batch_op.alter_column("discharge_coefficient_negative", nullable=True)
+        batch_op.alter_column("discharge_coefficient_positive", nullable=True)
 
 
 def upgrade():
@@ -84,9 +72,7 @@ def upgrade():
     if conn.dialect.name == "sqlite" and "spatial_ref_sys" not in existing_tables:
         op.execute("SELECT InitSpatialMetadata()")
 
-    # Check if migration 173 should be done
-    upgrade_173(inspector)
-
+    upgrade_173()
 
     create_table_if_not_exists(
         "v2_2d_boundary_conditions",